+++ application/views/templates/header.php
@@ -34,6 +34,9 @@
   <meta property="og:locale" content="hu_HU" />
   <meta property="og:site_name" content="Ekultura.hu" />
   <meta property="fb:app_id" content="454047391791470" />
+  <?php if (isset($canonical_url)): ?>
+    <link rel="canonical" href="<?php echo $canonical_url; ?>" />
+  <?php endif; ?>
   <link rel="icon" type="image/x-icon" href="http://ekultura.hu/favicon.ico">
   <link rel="alternate" href="<?php echo site_url('rss');?>" title="ekultura.hu RSS feed" type="application/rss+xml" />
   <link type="text/css" href="<?php echo base_url('assets/bootstrap/css/bootstrap-cosmo.css');?>" rel="stylesheet">