+++ application/logs/CHANGELOG.md
@@ -9,6 +9,7 @@
 - bugfix: felesleges "height: auto" törlése a cikközbeni képekről
 - bugfix: felesleges outline törlése a hamburger buttonről
 - a meták törlésekor jelenjen meg, hogy hány darab van belőle használatban
+- canonical (gyűjtő) oldal megadása a googlenek
 
 ## Változások 2018.12.09.